scripts/requirements.txt
requests>=2.28
pyserial>=3.5
